Move segyio/_gather.py -> segyio/gather.py
