+++ python/segyio/gather.py
