+++ python/segyio/segy.py
@@ -23,7 +23,7 @@ except (NameError, ImportError): pass
 import numpy as np
 
 from segyio.header import Header
-from segyio._gather import Gather
+from .gather import Gather
 from segyio._line import Line
 from segyio.trace import Trace
 from segyio._field import Field
